+++ Public/New-StringDecryption.ps1
@@ -1,22 +1,61 @@
 ﻿function New-StringDecryption
 {
+  <#
+      .SYNOPSIS
+      Cmdlet will decode a base64 encrypted string
+	
+      .DESCRIPTION
+      Function will take as input a base64 encrypted string and output the unencrypted string
+	
+      .PARAMETER EncryptedString
+      The string to be decrypted
+	
+      .PARAMETER EncryptPassPhrase
+      A string representing the encryption passphrase to be used to decrypt the string.
+		
+      If not specified hostname of the computer where function is invoked will be used.
+	
+      .PARAMETER EncryptSalt
+      A string representing the encryption salt to use to decrypt the string.
+		
+      If not specified hostname of the computer where function is invoked will be used.
+	
+      .PARAMETER IntersectingVector
+      A string representing the intersecting vector to be used during strict decryption.
+		
+      If not specified it will default to a standard value which must match the one used to encrypt the string.
+		
+      For better security this should be changed to a custom string eatch time.
+	
+      .EXAMPLE
+      PS C:\> New-StringDecryption
+	
+      .OUTPUTS
+      System.String
+  #>
+	
 	[OutputType([string])]
 	param
 	(
+		[Parameter(Mandatory = $true)]
+		[ValidateNotNullOrEmpty()]
 		[string]
 		$EncryptedString,
+		[ValidateNotNullOrEmpty()]
 		[string]
-		$EncryptPassPhrase,
+		$EncryptPassPhrase = $env:Computername,
+		[ValidateNotNullOrEmpty()]
 		[string]
-		$EncryptSalt,
+		$EncryptSalt = $env:Computername,
+		[ValidateNotNullOrEmpty()]
 		[string]
 		$IntersectingVector = 'Q!L@2QTCYgsG'
 	)
 	
 	# Instantiate empty return value
-	[string]$returnValue = $null
+	[string]$DecryptedString = $null
 	
-	# Reg
+	# Regex to check if string is in the correct format
 	[regex]$base64RegEx = '^([A-Za-z0-9+/]{4})*([A-Za-z0-9+/]{3}=|[A-Za-z0-9+/]{2}==)?$'
 	
 	# Instantiate COM Object for RijndaelManaged Cryptography 
@@ -44,22 +83,9 @@
 		return
 	}
 	
-	# Check if we have a passphrase
-	if ([string]::IsNullOrEmpty($EncryptPassPhrase) -eq $true)
-	{
-		$EncryptPassPhrase = $env:Computername
-	}
-	
-	# Check if we have a salt value
-	if ([string]::IsNullOrEmpty($EncryptSalt) -eq $true)
-	{
-		$EncryptSalt = $env:Computername
-	}
-	
 	# Convert Salt and Passphrase to UTF8 Bytes array
 	[System.Byte[]]$byteEncryptSalt = [Text.Encoding]::UTF8.GetBytes($EncryptSalt)
 	[System.Byte[]]$bytePassPhrase = [Text.Encoding]::UTF8.GetBytes($EncryptPassPhrase)
-	
 	
 	# Create the Encryption Key using the passphrase, salt and SHA1 algorithm at 256 bits
 	$encryptionObject.Key = (New-Object Security.Cryptography.PasswordDeriveBytes $bytePassPhrase,
@@ -71,7 +97,7 @@
 	$encryptionObject.IV = (New-Object Security.Cryptography.SHA1Managed).ComputeHash([Text.Encoding]::UTF8.GetBytes($IntersectingVector))[0 .. 15]
 	
 	# Create new decryptor Key and IV
-	$objectDecryptor = $encryptionObject.CreateDecryptor()
+	[System.Security.Cryptography.RijndaelManagedTransform]$objectDecryptor = $encryptionObject.CreateDecryptor()
 	
 	# Create a New memory stream with the encrypted value
 	[System.IO.MemoryStream]$memoryStream = New-Object IO.MemoryStream  @( ,$encryptedStringByte)
@@ -82,21 +108,44 @@
 	# Read the new decrypted stream 
 	[System.IO.StreamReader]$streamReader = New-Object IO.StreamReader $cryptoStream
 	
-	# Return from the function the stream 
-	$returnValue = $streamReader.ReadToEnd()
+	try
+	{
+		# Return from the function the stream 
+		[string]$DecryptedString = $streamReader.ReadToEnd()
+		
+		# Stop the stream     
+		$streamReader.Close()
+		
+		# Stop the crypto stream 
+		$cryptoStream.Close()
+		
+		# Stop the memory stream 
+		$memoryStream.Close()
+		
+		# Clears all crypto objects 
+		$encryptionObject.Clear()
+		
+		# Return decrypted string
+		return $DecryptedString
+	}
 	
-	# Stops the stream     
-	$streamReader.Close()
-	
-	# Stops the crypto stream 
-	$cryptoStream.Close()
-	
-	# Stops the memory stream 
-	$memoryStream.Close()
-	
-	# Clears all crypto objects 
-	$encryptionObject.Clear()
-	
-	# Return 
-	return $returnValue
+	catch
+	{
+		# Save exception
+		[string]$reportedException = $Error[0].Exception.Message
+		
+		Write-Warning -Message "String $EncryptedString could not be decripted - Use the -Verbose paramter for more details"
+		
+		# Check we have an exception message
+		if ([string]::IsNullOrEmpty($reportedException) -eq $false)
+		{
+			Write-Verbose -Message $reportedException
+		}
+		else
+		{
+			Write-Verbose -Message 'No inner exception reported by Disconnect-AzureAD cmdlet'
+		}
+		
+		return [string]::Empty
+	}	
 }